strategic_nifty_vision/requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
typing-extensions
